Fix compiler warnings (macOS + Linux)
Resolve the GCC/clang -Wundefined-inline ("inline used but never
defined") warnings that surfaced on the 64-bit ports, and a handful of
related correctness issues exposed while cleaning up the build.

Warning fixes
- Include the *.inline.hpp headers from each .cpp TU that uses the
  inline functions they define (oop, memOop, universe.store,
  generation, rSet). The definitions were previously reached only
  indirectly through universe.hpp, so the compiler saw the
  declarations but not their bodies.
- Fix generation.inline.hpp's stale include guard, which shared
  _GENERATION_HPP with generation.hpp and made it a no-op.
- Move integerOps' helper constants and inline bodies (as_Digit,
  axpy, ...) into its header so every TU sees them.

64-bit / pointer-width correctness (x86-64, arm64)
- nativeInstruction::data() => intptr_t
- eventLog.hpp LOG_EVENT* macros cast through (void*)(intptr_t)
- nmethod::isFree(), markOop::context_forward_bit_mask() casts

Portability / robustness
- os_darwin.cpp: sprintf -> snprintf
- code/ncode.hpp: abort() in the placeholder operator new

Tooling
- build.unix: add make format / format-check targets and .clang-format
- CI: add a clang-format check job (build-unix.yml)

+++ vm/code/nativeInstruction.hpp
@@ -119,8 +119,8 @@ class NativeMov: public NativeInstruction {
 
   char* instruction_address() const		{ return addr_at(instruction_offset); }
   char* next_instruction_address() const	{ return addr_at(next_instruction_offset); }
-  int   data() const				{ return long_at(data_offset); }
-  void  set_data(int x)				{ set_long_at(data_offset, x); }
+  intptr_t data() const				{ return long_at(data_offset); }
+  void  set_data(intptr_t x)			{ set_long_at(data_offset, x); }
 
   void  verify();
   void  print();
@@ -155,8 +155,8 @@ class NativeTest: public NativeInstruction {
 
   char* instruction_address() const		{ return addr_at(instruction_offset); }
   char* next_instruction_address() const	{ return addr_at(next_instruction_offset); }
-  int   data() const				{ return long_at(data_offset); }
-  void  set_data(int x)				{ set_long_at(data_offset, x); }
+  intptr_t data() const				{ return long_at(data_offset); }
+  void  set_data(intptr_t x)			{ set_long_at(data_offset, x); }
 
   void  verify();
   void  print();
